refact: 렌더링 역할 이전
