+++ src/main/java/baseball/domain/ball/GuessResult.java
@@ -1,0 +1,14 @@
+package baseball.domain.ball;
+
+public record GuessResult(
+        int ballCount,
+        int strikeCount
+) {
+    public boolean hasThreeStrike() {
+        return strikeCount == BallNumbers.BALL_COUNT;
+    }
+
+    public boolean hasNothing() {
+        return ballCount + strikeCount == 0;
+    }
+}
